Refactor lambda capture in MainWindow and improve file reading logic
- Updated the lambda in `MainWindow::MainWindow` to capture `this` implicitly, simplifying the code.
- Modified the `ReadAtOffset` function in `tailing_bytes_producer.cpp` to use a more explicit `off_t` type for the file offset, enhancing clarity and type safety.
- Adjusted test cases in `test_json_parser.cpp`, `test_log_table.cpp`, and `test_tailing_bytes_producer.cpp` to ensure proper type handling and improve code consistency.

These changes enhance code readability and maintainability across the project.

+++ test/lib/src/test_json_parser.cpp
@@ -661,7 +661,7 @@ TEST_CASE("Parallel parse parity vs. single-thread", "[json_parser][parity]")
     {
         glz::generic_sorted_u64 json;
         json["index"] = static_cast<int64_t>(i);
-        json["level"] = std::string(LEVELS[levelDist(rng)]);
+        json["level"] = std::string(LEVELS[static_cast<size_t>(levelDist(rng))]);
         json["component"] = std::string("component_") + std::to_string(i % 7);
         json["message"] = std::string("event ") + std::to_string(i) + " — value " + std::to_string(intDist(rng));
         json["counter"] = static_cast<int64_t>(intDist(rng));
